Initial stateless reset detection
Backport the stateless reset patch to AOSP (current
version of quiche is 0.14.0) so that we don't need to
wait until the patch is released to a quiche version
newer than 0.16.0 (currently, the latest version is 0.16.0).

This patch is slightly modified because the type of
stateless_reset_token is different (In 0.14.0 the
type is Option<Vec<u8>>; in 0.16.0 the type is
Option<u128>).

Source patch:
https://github.com/cloudflare/quiche/commit/c6357db0b5311010e266637eda2f645b7fa91df4.

Bug: 242832641
Bug: 245074765
Bug: 258767218
Test: cd packages/modules/DnsResolver && atest
Test: cd external/rust/crates/quiche && atest
Test: 1. Applied ag/20124672 to DnsResolver
      2. Ran dnschk every 5 minutes for 1 hour
      3. Checked the log:
        a. Confirmed that some stateless reset packets were received
        b. Confirmed that DNS queries fallback'ed to DoT immediately
           after DnsResolver received stateless reset packets
Change-Id: Ife933f54ac6ec1098a9046673ca200c6b4e2ebbf
(cherry picked from commit e1d5b663756184f5737c886a01cfa519ef256508)
Merged-In: Ife933f54ac6ec1098a9046673ca200c6b4e2ebbf